testproject/pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "alembic>=1.18.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.128.0",
    "greenlet>=3.3.0",
    "grpcio-tools>=1.68.0",
    "jsonschema>=4.26.0",
    "prometheus-client>=0.21.1",
    "protobuf>=6.33.3",
    "pydantic-settings>=2.12.0",
    "python-keycloak>=7.0.2",
    "redis>=7.1.0",
    "sqlmodel>=0.0.31",
    "urllib3>=2.6.3",
    "uvicorn>=0.40.0",
    "websockets>=16.0",
]

[dependency-groups]
dev = [
    "bandit>=1.9.1",
    "black>=24.0.0",
    "interrogate>=1.7.0",
    "ipython>=9.9.0",
    "jinja2>=3.1.6",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "pytest-xdist>=3.8.0",
    "rich>=13.0.0",
    "ruff>=0.14.11",
    "skjold>=0.6.2",
    "testcontainers>=4.9.0",
    "typer>=0.21.1",
    "typos>=1.42.0",
    "vulture>=2.14",
]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.26.0",
    "mkdocs-git-revision-date-localized-plugin>=1.3.0",
    "mkdocs-minify-plugin>=0.8.0",
]
profiling = [
    "scalene>=1.5.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
norecursedirs = [".worktree", ".git", ".venv", "__pycache__"]
markers = [
    "integration: marks tests as integration tests (require external services like Keycloak)",
    "load: marks tests as load tests (high resource usage, skip by default)",
    "chaos: marks tests as chaos engineering tests (simulate failures)",
]

[tool.coverage.run]
omit = [
    "*__init__*",
    "*__main__*",
    "*/tests/*",
    "app/settings.py",
    "app/storage/migrations/*",
    "app/protocols.py",
    "app/schemas/*"
]

[tool.coverage.report]
show_missing = true
fail_under = 81  # Minimum coverage threshold (increased from 80%)
exclude_lines = [
    "pragma: no cover",
    "from",
    "import",
    "pass",
    "if __name__ == .__main__.:",
]

[tool.coverage.html]
directory = "htmlcov"

[tool.vulture]
ignore_names = [
    "create_author_endpoint",
    "get_authors_endpoint",
    "get_paginated_authors_endpoint",
    "Web",
    "dispatch",
    "get_authors_handler",
    "get_paginated_authors_handler",
    "authenticate",
    "dialect",  # SQLAlchemy type decorator required parameter
    "executemany"  # SQLAlchemy event listener required parameter
]
# exclude = ["app/api/ws/constants.py"]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    ".venv",
    ".worktree"
]
line-length = 79
indent-width = 4
target-version = "py313"


[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
# ignore = ["F401", "F811", "F821"]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.interrogate]
ignore-module = true
ignore-init-method = true
ignore-nested-functions = true
ignore-property-decorators = true
fail-under = 80
verbose=2
ignore-regex = ["__new__"]
exclude = [
    ".worktree",
    "tests",
    "scripts",
    "app/settings.py",
    "app/logging.py",
    "app/schemas/*",
    "app/storage/migrations/versions/*"
]

[tool.mypy]
exclude = [
    "app/schemas/proto/",  # Exclude auto-generated protobuf files
]
disable_error_code = ["untyped-decorator", "no-any-return"]

[[tool.mypy.overrides]]
module = "app.schemas.proto.*"
ignore_errors = true

[tool.skjold]
sources = ["github", "pyup", "gemnasium"]  # Sources to check against.
report-only = true                         # Report only, always exit with zero.
report-format = 'json'                     # Output findings as `json`. Default is 'cli'.
cache_dir = '.skjold_cache'                # Cache location (default: `~/.skjold/cache`).
cache_expires = 86400                      # Cache max. age.
verbose = true
